Move LUA functions to library
